Fixed previous issue with UI.py
Issue where we couldn't clear the tree list before adding new games to it.

Issue found was that if the search found a redeemed game it would count this as a result, even though we don't create a treeview item for it. Therefore when the program went to delete by id, it wouldn't be able to find anything to delete.

Solved by instead asking Treeview for a  tuple of all top level children (currently all games will be top level) and then checking the length of these to run the list clearing operation

I hope that made sense, it probably didn't it's gone 11pm