Double click to copy and set game to redeemed
Added functionality so user can double click on a game in the tree to copy it to keyboard, this also sets the games Redeemed status to 1, so it will no longer show up in searches.

I wanted to make this a manual operation with a checkbox, however tkinter's treeview doesn't let you insert UI widgets into it's columns. Considering moving to QT due to the number of limitations I have run into with Tkinter.